rep/fastflow/examples/requirements_common.txt
# Common
matplotlib==3.5.3
pandas
numpy==1.20.3
protobuf==3.19.6
tensorflow-metadata==1.12.0
tensorflow_datasets==4.8.0
tensorflow_io==0.30.0
tensorflow_addons
imgaug
